add screenshots
correct typo in instruction message

add screenshots

+++ order-service/src/main/kotlin/OrderService.kt
@@ -2,7 +2,7 @@ package com.amex.fruitstand.order
 
 import java.math.BigInteger
 
-data class OrderSubmissionResult(val orderPrice: BigInteger?, val errorMessage: String?)
+data class OrderSubmissionResult(val orderPrice: Int?, val errorMessage: String?)
 
 interface OrderService {
     fun processOrder(orderInput: String): OrderSubmissionResult
